+++ packages/shared/src/react/hooks/useSession.ts
@@ -1,11 +1,6 @@
-import type { ActiveSessionResource } from '@clerk/types';
+import type { UseSessionReturn } from '@clerk/types';
 
 import { useAssertWrappedByClerkProvider, useSessionContext } from '../contexts';
-
-type UseSessionReturn =
-  | { isLoaded: false; isSignedIn: undefined; session: undefined }
-  | { isLoaded: true; isSignedIn: false; session: null }
-  | { isLoaded: true; isSignedIn: true; session: ActiveSessionResource };
 
 type UseSession = () => UseSessionReturn;
